add lot alert and error message

+++ P5-Dev-Web-Kanap-master/front/js/cart.js
@@ -1,21 +1,36 @@
 
 let listLocalStorage = JSON.parse(localStorage.getItem('objetL'));
-lengthlistLocalStorage = listLocalStorage.length;
 
 let total = 0;
 let totalQuantity = 0;
+function viewCartEmpty(){
+  const h1 = document.getElementsByTagName('h1');
+  const aH1 = document.createElement('a')
+  aH1.textContent = 'ici';
+  h1[0].textContent = `Votre panier est vide, vous trouverez nos super canapé,`;
+  aH1.setAttribute('href', 'index.html')
+  h1[0].appendChild(aH1);
+}
+//if Cart empty 
+if (listLocalStorage) {
+  lengthlistLocalStorage = listLocalStorage.length;
+    if (lengthlistLocalStorage == 0) {
+    viewCartEmpty()
+  }
+}
+else {
+  viewCartEmpty()
+}
 
 for (let i in listLocalStorage) {
-
   async function fetchlistLocalStorageJSON() {
     const response = await fetch("http://localhost:3000/api/products/" + listLocalStorage[i].idProduct);
     const product = await response.json();
     return product;
   }
+
   fetchlistLocalStorageJSON().then(product => {
-
     // creat div
-
     var article = document.createElement('article');
     const divCartItemImg = document.createElement('div');
     var img = document.createElement('img');
@@ -88,27 +103,36 @@ for (let i in listLocalStorage) {
     nodeTotalquantity.textContent = totalQuantity;
     //suppression product cart
     pDeleteItem.addEventListener('click', () => {
-      listLocalStorage.splice(i, 1);
-      localStorage.setItem("objetL", JSON.stringify(listLocalStorage));
-      document.location.reload(true);
+      if (confirm("Etes-vous sûre de vouloir supprimer cet article?")) {
+        listLocalStorage.splice(i, 1);
+        localStorage.setItem("objetL", JSON.stringify(listLocalStorage));
+        document.location.reload(true);
+      }
     })
     //modifictaion quantity input
     inputQuantity.onchange = function () {
-      var differencyQuantity = this.value - listLocalStorage[i].quantityItem;
-      //change quantity in input
-      listLocalStorage[i].quantityItem = this.value;
-      localStorage.setItem("objetL", JSON.stringify(listLocalStorage));
-      inputQuantity.setAttribute('value', listLocalStorage[i].quantityItem);
-      //change price
-      var oldPrice = pricebyvanap;
-      pricebyvanap = Number(listLocalStorage[i].quantityItem) * product.price;
-      pPrice.textContent = pricebyvanap + " €";
-      //change quantity
-      totalQuantity = totalQuantity + differencyQuantity;
-      nodeTotalquantity.textContent = totalQuantity;
-      //change total price
-      total = total + pricebyvanap - oldPrice;
-      nodeTotalPrice.textContent = total;
+      var bolRegexQuantity = (/^[1-9]$|^[1-9][0-9]$|^(100)$/.test(inputQuantity.value));
+      if (bolRegexQuantity == false) {
+        alert('Veuillez saisir une quantité entre 1 et 100');
+        inputQuantity.value = "listLocalStorage[i].quantityItem";
+      }
+      else {
+        var differencyQuantity = this.value - listLocalStorage[i].quantityItem;
+        //change quantity in input
+        listLocalStorage[i].quantityItem = this.value;
+        localStorage.setItem("objetL", JSON.stringify(listLocalStorage));
+        inputQuantity.setAttribute('value', listLocalStorage[i].quantityItem);
+        //change price
+        var oldPrice = pricebyvanap;
+        pricebyvanap = Number(listLocalStorage[i].quantityItem) * product.price;
+        pPrice.textContent = pricebyvanap + " €";
+        //change quantity
+        totalQuantity = totalQuantity + differencyQuantity;
+        nodeTotalquantity.textContent = totalQuantity;
+        //change total price
+        total = total + pricebyvanap - oldPrice;
+        nodeTotalPrice.textContent = total;
+      }
     }
   })
 
@@ -117,6 +141,7 @@ for (let i in listLocalStorage) {
 // form contact 
 const contact = document.getElementById('order');
 contact.addEventListener('click', function () {
+
   event.preventDefault();
   var contactTab = {
     firstName: document.getElementById('firstName').value,
@@ -133,41 +158,20 @@ contact.addEventListener('click', function () {
   var bolcity = /^[a-zA-Z0-9\u00C0-\u00FF\s,. '-]{3,}$/.test(contactTab.city);
   // validate firstName
 
-  if (bolFirstName) {
-    firstNameErrorMsg.innerHTML = "";
-  } else {
-    firstNameErrorMsg.textContent = "ceci n\'est pas un prénom";
-    event.preventDefault();
+  // verification if quantity between 1 and 100
+  function verificateFormatInput(bolInput, idErrorMessage, Message){
+    if (bolInput) {
+      idErrorMessage.innerHTML = "";
+    } else {
+      idErrorMessage.textContent = Message;
+      event.preventDefault();
+    }
   }
-  //validate email
-  if (bolemail) {
-    emailErrorMsg.innerHTML = "";
-  } else {
-    emailErrorMsg.textContent = "ceci n\'est pas un email";
-    event.preventDefault();
-  }
-  //validate lastName
-  if (bollastName) {
-    lastNameErrorMsg.innerHTML = "";
-  } else {
-    lastNameErrorMsg.textContent = "ceci n\'est pas un nom";
-    event.preventDefault();
-  }
-  // validate adress
-  if (boladdress) {
-    addressErrorMsg.innerHTML = "";
-  } else {
-    addressErrorMsg.textContent = "ceci n'est pas une adresse";
-    event.preventDefault();
-  }
-  //validate city
-  if (bolcity) {
-    cityErrorMsg.innerHTML = "";
-  } else {
-    cityErrorMsg.textContent = "ceci n'est pas une ville";
-    event.preventDefault()
-  }
-
+  verificateFormatInput(bolFirstName, firstNameErrorMsg, "ceci n\'est pas un prénom");
+  verificateFormatInput(bolemail, emailErrorMsg, "ceci n\'est pas un email");
+  verificateFormatInput(bollastName, lastNameErrorMsg, "ceci n\'est pas un nom");
+  verificateFormatInput(boladdress, addressErrorMsg, "ceci n'est pas une adresse");
+  verificateFormatInput(bolcity, cityErrorMsg, "ceci n'est pas une ville");
 
 
   var contact = contactTab;
